Add `avoidEarlyExit` in setting
The purpose of this setting is to avoid to draw the card with the exit
too soon. Indeed, if draw too soon, it could be possible to the players
to be in a situation where they are stuck and can't complete the game.
Before this commit, they can put the exit card into the deck with the
dedicated button, but with this setting, it easier as the user doesn't
have to manage such situations.